+++ CHANGELOG.md
@@ -22,3 +22,4 @@
 Bug fixes and increase customizations:
 * Fix a bug that cause outermost card clipped inside widget's boundaries
 * Allow user to disable / enable carousel's swipe interaction
+* Allow user to disable automatic loop